requirements.txt
python-docx==1.1.0
PyYAML==6.0.1
pandas==2.1.4
click==8.1.7
jinja2==3.1.2
openpyxl==3.1.2
colorama==0.4.6
pywin32==306
